27 sep 2023 update; add loss curve to NN; refactor makefiles and scripts; testing updated eisvogel latex template for pandoc 3.1.8;

+++ python/visualize-other.py
@@ -20,47 +20,49 @@ if os.path.exists("assets/data"):
     # Visualize Clapeyron slopes for 660 transition
     visualize_training_dataset_design(Pmin, Pmax, Tmin, Tmax)
 
-    # Visualize benchmark computation times
-    visualize_benchmark_efficiency(figdir, "benchmark-efficiency.png")
+    if os.path.exists(f"assets/data/benchmark-efficiency.csv"):
+        # Visualize benchmark computation times
+        visualize_benchmark_efficiency(figdir, "benchmark-efficiency.png")
 
-    for t in targets:
-        # Visualize rocml performance metrics
-        visualize_rocml_performance(sampleid, t, res, figdir, "rocml")
+    if os.path.exists(f"assets/data/rocml-performance.csv"):
+        for t in targets:
+            # Visualize rocml performance metrics
+            visualize_rocml_performance(sampleid, t, res, figdir, "rocml")
 
-        # First row
-        combine_plots_horizontally(
-            f"figs/rocml-inference-time-mean-{sampleid}-{res}.png",
-            f"figs/rocml-training-time-mean-{sampleid}-{res}.png",
-            "figs/temp1.png",
-            caption1="a)",
-            caption2="b)"
-        )
+            # First row
+            combine_plots_horizontally(
+                f"figs/rocml-inference-time-mean-{sampleid}-{res}.png",
+                f"figs/rocml-training-time-mean-{sampleid}-{res}.png",
+                "figs/temp1.png",
+                caption1="a)",
+                caption2="b)"
+            )
 
-        os.remove(f"figs/rocml-inference-time-mean-{sampleid}-{res}.png")
-        os.remove(f"figs/rocml-training-time-mean-{sampleid}-{res}.png")
+            os.remove(f"figs/rocml-inference-time-mean-{sampleid}-{res}.png")
+            os.remove(f"figs/rocml-training-time-mean-{sampleid}-{res}.png")
 
-        # Second row
-        combine_plots_horizontally(
-            f"figs/rocml-rmse-test-mean-{t.replace('_', '-')}-{sampleid}-{res}.png",
-            f"figs/rocml-rmse-val-mean-{t.replace('_', '-')}-{sampleid}-{res}.png",
-            "figs/temp2.png",
-            caption1="c)",
-            caption2="d)"
-        )
+            # Second row
+            combine_plots_horizontally(
+                f"figs/rocml-rmse-test-mean-{t.replace('_', '-')}-{sampleid}-{res}.png",
+                f"figs/rocml-rmse-val-mean-{t.replace('_', '-')}-{sampleid}-{res}.png",
+                "figs/temp2.png",
+                caption1="c)",
+                caption2="d)"
+            )
 
-        os.remove(f"figs/rocml-rmse-test-mean-{t.replace('_', '-')}-{sampleid}-{res}.png")
-        os.remove(f"figs/rocml-rmse-val-mean-{t.replace('_', '-')}-{sampleid}-{res}.png")
+            os.remove(f"figs/rocml-rmse-test-mean-{t.replace('_', '-')}-{sampleid}-{res}.png")
+            os.remove(f"figs/rocml-rmse-val-mean-{t.replace('_', '-')}-{sampleid}-{res}.png")
 
-        # Stack rows
-        combine_plots_vertically(
-            "figs/temp1.png",
-            "figs/temp2.png",
-            f"figs/benchmark-rocmls-metrics-{t.replace('_', '-')}.png",
-            caption1="",
-            caption2=""
-        )
+            # Stack rows
+            combine_plots_vertically(
+                "figs/temp1.png",
+                "figs/temp2.png",
+                f"figs/rocml-performance-{t.replace('_', '-')}.png",
+                caption1="",
+                caption2=""
+            )
 
-        os.remove("figs/temp1.png")
-        os.remove("figs/temp2.png")
+            os.remove("figs/temp1.png")
+            os.remove("figs/temp2.png")
 
 print("visualize-other.py done!")
